// api-stage/server/src/modules/code-template/code-template.enums.ts
export enum ModuleType {
    WAREHOUSE = 'WAREHOUSE',
    CASHBOX = 'CASHBOX',
    PERSONNEL = 'PERSONNEL',
    PRODUCT = 'PRODUCT',
    CUSTOMER = 'CUSTOMER',
    INVOICE_SALES = 'INVOICE_SALES',
    INVOICE_PURCHASE = 'INVOICE_PURCHASE',
    ORDER_SALES = 'ORDER_SALES',
    ORDER_PURCHASE = 'ORDER_PURCHASE',
    INVENTORY_COUNT = 'INVENTORY_COUNT',
    QUOTE = 'QUOTE',
    DELIVERY_NOTE_SALES = 'DELIVERY_NOTE_SALES',
    DELIVERY_NOTE_PURCHASE = 'DELIVERY_NOTE_PURCHASE',
    WAREHOUSE_TRANSFER = 'WAREHOUSE_TRANSFER',
    TECHNICIAN = 'TECHNICIAN',
    WORK_ORDER = 'WORK_ORDER',
    SERVICE_INVOICE = 'SERVICE_INVOICE',
    POS_CONSOLE = 'POS_CONSOLE',
    CHECK_BILL_JOURNAL = 'CHECK_BILL_JOURNAL',
    CHECK_BILL_DOCUMENT = 'CHECK_BILL_DOCUMENT',
}